requirements_shfe.txt
# Flask and Web Server
flask==2.3.3
gunicorn==21.2.0

# Selenium for Web Scraping
selenium==4.15.2

# LLM Integration
google-generativeai==0.5.4

# Data Export
xlwt==1.3.0

# Google Cloud Integration
google-cloud-storage==2.10.0